app/requirements.txt
mcp[cli]>=1.0
openai>=1.0
fastapi>=0.115
uvicorn[standard]>=0.32
python-multipart>=0.0.9
python-dotenv>=1.0
